ai_worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai_worker"
version = "0.1.0"
description = "Multi-Agent AI Employee System"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "discord.py>=2.3.0",
    "lark-oapi>=1.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
ai-worker = "ai_worker.main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
